Task_2

+++ Task_2.py
@@ -1,8 +1,8 @@
 # 2. The Greatest Showdown
 # Task 2: Identify the Smallest
-num1 = float(input("Enter the first number: "))
-num2 = float(input("Enter the second number: "))
-num3 = float(input("Enter the third number: "))
+num1 = int(input("Enter the first number: "))
+num2 = int(input("Enter the second number: "))
+num3 = int(input("Enter the third number: "))
 
 if num1 >= num2 and num1 >= num3:
     largest = num1
